fix NoteViewController keyboard, move to NotesView to NoteViewController.
