+++ MyBujo2.0/Use Cases/Daily/Drawing/Controllers/DrawingViewController.swift
@@ -31,9 +31,6 @@ class DrawingViewController: MediaViewController {
 		self.title = "Drawing"
     }
 	
-	
-	
-
     override func viewDidLoad() {
         super.viewDidLoad()
         canvasWidth = contentView.frame.width
