+++ MyBujo2.0/Use Cases/Daily/Goals/Views/GoalTableViewCell.swift
@@ -18,8 +18,7 @@ class GoalTableViewCell: UITableViewCell, ViewCode {
         lbl.numberOfLines = 0
         lbl.font = UIFont(name: "AvenirNext-Medium", size: 16)
         lbl.textColor = UIColor(named: "TitleColor")!
-		lbl.backgroundColor = .red
-        return lbl
+		return lbl
     }()
     
     let goalBullet: UIImageView = {
